requirements.txt
PyQt6
Pillow
pillow-avif-plugin
imagecodecs
cx_Freeze